# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartShelf API",
        "description": "API documentation for SmartShelf project",
        "version": "1.0.0"
    },
    "paths": {
        "/api/produits": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Lister tous les produits",
                "operationId": "059ba901d56d1b85845397dcc5120bd2",
                "responses": {
                    "200": {
                        "description": "Liste des produits",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Produit"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Produits"
                ],
                "summary": "Créer un nouveau produit",
                "operationId": "1d2c1c7cee2e131dfeb85184f2824a90",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Produit"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produit créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produit"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/produits/{id}": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Afficher un produit spécifique",
                "operationId": "4f89d6d658ffba7d843303b681d52762",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du produit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produit"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Produits"
                ],
                "summary": "Mettre à jour un produit",
                "operationId": "05d27b4688f87d159b38ecb2f0d03dc1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Produit"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produit"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Produits"
                ],
                "summary": "Supprimer un produit",
                "operationId": "3a0611f76b7e02b393d422a5598e5afd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produit supprimé avec succès"
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "properties": {
                    "titre": {
                        "description": "Titre de la catégorie",
                        "type": "string",
                        "example": "Électronique"
                    },
                    "description": {
                        "description": "Description de la catégorie",
                        "type": "string",
                        "example": "Catégorie pour tous les produits électroniques."
                    }
                },
                "type": "object"
            },
            "Produit": {
                "title": "Produit",
                "properties": {
                    "nom": {
                        "description": "Nom du produit",
                        "type": "string",
                        "example": "Produit A"
                    },
                    "description": {
                        "description": "Description du produit",
                        "type": "string",
                        "example": "Ceci est un produit de haute qualité."
                    },
                    "prix": {
                        "description": "Prix du produit",
                        "type": "number",
                        "format": "float",
                        "example": 19.99
                    },
                    "quantite": {
                        "description": "Quantité du produit",
                        "type": "integer",
                        "example": 100
                    },
                    "categorie_id": {
                        "description": "ID de la catégorie du produit",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "Rayon": {
                "title": "Rayon",
                "properties": {
                    "nom": {
                        "description": "Nom du rayon",
                        "type": "string",
                        "example": "Rayon A"
                    },
                    "description": {
                        "description": "Description du rayon",
                        "type": "string",
                        "example": "Rayon pour les produits en promotion."
                    },
                    "category_id": {
                        "description": "ID de la catégorie du rayon",
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Produits",
            "description": "Gestion des produits"
        }
    ]
}